地址标签CustomerAddressLabel labelType的类型修改

+++ service-projects/customer-service/src/main/java/com/winhxd/b2c/customer/service/impl/CustomerAddressServiceImpl.java
@@ -113,7 +113,7 @@ public class CustomerAddressServiceImpl implements CustomerAddressService {
         CustomerAddressLabel customerAddressLabel = new CustomerAddressLabel();
         BeanUtils.copyProperties(customerAddressLabelCondition, customerAddressLabel);
         customerAddressLabel.setCreated(new Date());
-        customerAddressLabel.setLabelType(2);
+        customerAddressLabel.setLabelType((short) 2);
         return customerAddressLabelMapper.insertSelective(customerAddressLabel);
     }
